requirements.txt
asgiref==3.10.0
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cssselect2==0.8.0
dj-database-url==2.3.0
Django==5.2.8
djangorestframework==3.15.2
et_xmlfile==2.0.0
Flask==3.1.2
fonttools==4.60.1
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.0
MarkupSafe==3.0.3
mssql-django==1.6
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
pyodbc==5.3.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.11.0
zopfli==0.4.0
